ini: Enable gMulticastHostFwMsgs for kona, lahaina and default
gMulticastHostFwMsgs should be enabled for sending of diag_event
and diag_log_packet to the userspace through netlink. There are
various diag_events which are sent to userspace and one of them
is data stall event. Without enabling this the data stall event
will not get routed and would impact the data stall detection
framework.

CRs-Fixed: 2614392
Change-Id: Ia9c6edf69e7bb1d6736c82ed0c5cb548b715e8b1