wlan: Disable enhanced GreenAP for Hasting
This feature is not required for auto SP,
so disable it. Once enabled this feature,
it will enter EGAP mode while there is no
STAs connected or all STAs fall sleep,
and then SAP will force setting chainmask
to 0x1, which mean it will use chain0 only
to save the power. With this case, if the
hw is without chain0 connected, then such
feature will block beacon broadcasting.

Change-Id: Id34eefa435ffb01368592495d34dd3122946a98f
Signed-off-by: Chaoli Zhou <zchaoli@codeaurora.org>