wlan: Reduce ipa tx buffer count
Since wlan fw share the tx desc count for both
sw and hw path, which may be not enough for the
sw and hw concurrency test case if hw path consume
most of the tx desc counts. So limit the ipa hw
tx desc count to mitigate the issue.
With this change, peak KPI is not impacted.

Change-Id: I8ee57a5dabd5e0e069e88d5910e232baaf9aeda1